# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Contacts API",
      "description": "A REST API for managing contacts stored in MongoDB.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://cse341-contacts-318h.onrender.com",
        "description": "Production (Render)"
      },
      {
        "url": "http://localhost:8080",
        "description": "Local development"
      }
    ],
    "paths": {
      "/contacts": {
        "get": {
          "summary": "Get all contacts",
          "description": "Returns a list of all contacts in the database.",
          "responses": {
            "200": {
              "description": "A list of contacts",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Contact" }
                  }
                }
              }
            },
            "500": { "description": "Internal server error" }
          }
        },
        "post": {
          "summary": "Create a new contact",
          "description": "Creates a new contact. All fields are required.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContactInput" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Contact created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "64a1f2c3e4b0a12345678901"
                      }
                    }
                  }
                }
              }
            },
            "400": { "description": "Missing required fields" },
            "500": { "description": "Internal server error" }
          }
        }
      },
      "/contacts/{id}": {
        "get": {
          "summary": "Get a contact by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" },
              "description": "MongoDB ObjectId of the contact"
            }
          ],
          "responses": {
            "200": {
              "description": "A single contact",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Contact" }
                }
              }
            },
            "400": { "description": "Invalid ID format" },
            "404": { "description": "Contact not found" },
            "500": { "description": "Internal server error" }
          }
        },
        "put": {
          "summary": "Update a contact by ID",
          "description": "Replaces a contact's data. All fields are required.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" },
              "description": "MongoDB ObjectId of the contact to update"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContactInput" }
              }
            }
          },
          "responses": {
            "204": { "description": "Contact updated successfully" },
            "400": { "description": "Invalid ID or missing fields" },
            "404": { "description": "Contact not found" },
            "500": { "description": "Internal server error" }
          }
        },
        "delete": {
          "summary": "Delete a contact by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" },
              "description": "MongoDB ObjectId of the contact to delete"
            }
          ],
          "responses": {
            "200": { "description": "Contact deleted successfully" },
            "400": { "description": "Invalid ID format" },
            "404": { "description": "Contact not found" },
            "500": { "description": "Internal server error" }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Contact": {
          "type": "object",
          "properties": {
            "_id": { "type": "string", "example": "64a1f2c3e4b0a12345678901" },
            "firstName": { "type": "string", "example": "Jane" },
            "lastName": { "type": "string", "example": "Doe" },
            "email": { "type": "string", "example": "jane.doe@example.com" },
            "favoriteColor": { "type": "string", "example": "blue" },
            "birthday": { "type": "string", "example": "1990-06-15" }
          }
        },
        "ContactInput": {
          "type": "object",
          "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"],
          "properties": {
            "firstName": { "type": "string", "example": "Jane" },
            "lastName": { "type": "string", "example": "Doe" },
            "email": { "type": "string", "example": "jane.doe@example.com" },
            "favoriteColor": { "type": "string", "example": "blue" },
            "birthday": { "type": "string", "example": "1990-06-15" }
          }
        }
      }
    }
  }